+++ src/main/kotlin/com/github/copyinaction/auth/service/CustomUserDetails.kt
@@ -1,0 +1,11 @@
+package com.github.copyinaction.auth.service
+
+import org.springframework.security.core.GrantedAuthority
+import org.springframework.security.core.userdetails.User
+
+class CustomUserDetails(
+    val id: Long,
+    username: String,
+    password: String,
+    authorities: Collection<GrantedAuthority>
+) : User(username, password, authorities)