+++ src/main/kotlin/com/github/copyinaction/booking/domain/Booking.kt
@@ -17,8 +17,9 @@ import java.util.*
 class Booking(
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
+    @Column(columnDefinition = "uuid")
     @Comment("예매 ID")
-    val id: UUID = UUID.randomUUID(),
+    val id: UUID? = null,
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "user_id")